Move functions away from utils to respective files
